requirements.txt
discord.py
quart
better-ipc
prisma
quart-discord
python-dotenv